+++ Lib/gftools/templates/text.html
@@ -5,11 +5,11 @@
 <b>Text</b>
 {% endblock %}
 {% block content %}
-  {% for font_class in font_classes %}
+  {% for font_class in css_font_classes or css_font_classes_before %}
       <div class="box">
 	<div class="box-title">{{ font_class.selector }} {{ pt_size }}pt</div>
 	<span class="{{ font_class.selector }}" style="font-size: {{ pt_size }}pt">
-	  {{ text }}
+	  {{ sample_text }}
           </span>
        </div>
     {% endfor %}
