+++ Lib/gftools/templates/waterfall.html
@@ -5,7 +5,7 @@
 <b>Waterfall</b>
 {% endblock %}
 {% block content %}
-  {% for font_class in font_classes %}
+  {% for font_class in css_font_classes or css_font_classes_before %}
     {% for pt_size in [7, 10, 11, 12, 14, 16, 18, 21, 27, 32] %}
       <div class="box">
 	<div class="box-title">{{ font_class.selector }} {{ pt_size }}pt</div>
